moving methods into separate categories
