+++ lib/ruby_outlook/apis/contacts.rb
@@ -1,0 +1,97 @@
+module RubyOutlook
+  class Client
+
+    # TODO - fix
+    # token (string): access token
+    # view_size (int): maximum number of results
+    # page (int): What page to fetch (multiple of view size)
+    # fields (array): An array of field names to include in results
+    # sort (hash): { sort_on => field_to_sort_on, sort_order => 'ASC' | 'DESC' }
+    # user (string): The user to make the call for. If nil, use the 'Me' constant.
+    def get_contacts(token, view_size, page, fields = nil, sort = nil, user = nil)
+      request_url = "/api/v2.0/" << (user.nil? ? "Me" : ("users/" << user)) << "/Contacts"
+      request_params = {
+        '$top' => view_size,
+        '$skip' => (page - 1) * view_size
+      }
+
+      unless fields.nil?
+        request_params['$select'] = fields.join(',')
+      end
+
+      unless sort.nil?
+        request_params['$orderby'] = sort[:sort_field] + " " + sort[:sort_order]
+      end
+
+      get_contacts_response = make_api_call "GET", request_url, token, request_params
+
+      JSON.parse(get_contacts_response)
+    end
+
+    # TODO - fix
+    # token (string): access token
+    # id (string): The Id of the contact to retrieve
+    # fields (array): An array of field names to include in results
+    # user (string): The user to make the call for. If nil, use the 'Me' constant.
+    def get_contact_by_id(token, id, fields = nil, user = nil)
+      request_url = "/api/v2.0/" << (user.nil? ? "Me" : ("users/" << user)) << "/Contacts/" << id
+      request_params = nil
+
+      unless fields.nil?
+        request_params = { '$select' => fields.join(',') }
+      end
+
+      get_contact_response = make_api_call "GET", request_url, token, request_params
+
+      JSON.parse(get_contact_response)
+    end
+
+    # TODO - fix
+    # token (string): access token
+    # payload (hash): a JSON hash representing the contact entity
+    # folder_id (string): The Id of the contact folder to create the contact in.
+    #                     If nil, contact is created in the default contacts folder.
+    # user (string): The user to make the call for. If nil, use the 'Me' constant.
+    def create_contact(token, payload, folder_id = nil, user = nil)
+      request_url = "/api/v2.0/" << (user.nil? ? "Me" : ("users/" << user))
+
+      unless folder_id.nil?
+        request_url << "/ContactFolders/" << folder_id
+      end
+
+      request_url << "/Contacts"
+
+      create_contact_response = make_api_call "POST", request_url, token, nil, nil, payload
+
+      JSON.parse(create_contact_response)
+    end
+
+    # TODO - fix
+    # token (string): access token
+    # payload (hash): a JSON hash representing the updated contact fields
+    # id (string): The Id of the contact to update.
+    # user (string): The user to make the call for. If nil, use the 'Me' constant.
+    def update_contact(token, payload, id, user = nil)
+      request_url = "/api/v2.0/" << (user.nil? ? "Me" : ("users/" << user)) << "/Contacts/" << id
+
+      update_contact_response = make_api_call "PATCH", request_url, token, nil, nil, payload
+
+      JSON.parse(update_contact_response)
+    end
+
+    # TODO - fix
+    # token (string): access token
+    # id (string): The Id of the contact to delete.
+    # user (string): The user to make the call for. If nil, use the 'Me' constant.
+    def delete_contact(token, id, user = nil)
+      request_url = "/api/v2.0/" << (user.nil? ? "Me" : ("users/" << user)) << "/Contacts/" << id
+
+      delete_response = make_api_call "DELETE", request_url, token
+
+      return nil if delete_response.nil? || delete_response.empty?
+
+       JSON.parse(delete_response)
+    end
+  
+  end
+end
